requirements.txt
pandas==2.0.3
Pillow==9.4.0
Pillow==10.2.0
scikit_learn==1.2.2
torch==2.1.1+cu121
torchvision==0.16.1+cu121
